闲置连接并完善登录流程

+++ common/config.go
@@ -41,5 +41,5 @@ type ConfigRedis struct {
 
 // ConfigRole : 配置 role 相关节
 type ConfigRole struct {
-	IdleTime int `default:"300" desc:"角色闲置该时间间隔后账号、角色对象内存中清除。单位：秒"`
+	IdleTime int64 `default:"300" desc:"角色闲置该时间间隔后账号、角色对象内存中清除。单位：秒"`
 }
